requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
botocore==1.27.0
certifi==2022.5.18.1
distlib==0.3.4
Django==4.0.4
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11
django-bootstrap4==22.1
django-storages==1.12.3
filelock==3.7.1
jmespath==1.0.0
Pillow==9.1.0
pipenv==2022.5.2
platformdirs==2.5.2
python-dateutil==2.8.2
pytube==12.0.0
pytz==2021.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
whitenoise==6.1.0
